Created CodeActionUtil (it doesn't do much yet)
